app/requirements.txt
fastapi==0.85.0
python-multipart==0.0.5 
python-dotenv==0.21.0
uvicorn==0.18.3
isodate==0.6.1 
requests==2.28.1

